speed up drawer tests 10s to 5s. added keypress action to screenshot tests
