+++ packages/dnb-design-system-portal/src/docs/uilib/components/drawer/demos.mdx
@@ -10,6 +10,7 @@ import {
   DrawerCallbackExample,
   DrawerCustomTriggerExample,
   DrawerNoAnimationNoSpacing,
+  DrawerScrollViewSetup,
 } from 'Docs/uilib/components/drawer/Examples'
 
 ## Demos
@@ -41,3 +42,7 @@ import {
 ### Remove animation and spacing
 
 <DrawerNoAnimationNoSpacing />
+
+<VisibleWhenVisualTest>
+  <DrawerScrollViewSetup />
+</VisibleWhenVisualTest>
