--- packages/dnb-design-system-portal/src/docs/uilib/components/drawer/visual-tests/hidden-tests.mdx
@@ -1,7 +1,0 @@
----
-draft: true
----
-
-import { DrawerScrollViewSetup } from 'Docs/uilib/components/drawer/visual-tests/Examples'
-
-<DrawerScrollViewSetup />
