+++ packages/dnb-eufemia/src/components/dialog/__tests__/Dialog.screenshot.test.ts
@@ -46,12 +46,7 @@ describe.each(['ui', 'sbanken'])('Dialog for %s', (themeName) => {
       simulate: 'click',
       simulateSelector:
         '[data-visual-test="dialog-help-button"] button:first-of-type',
-      simulateAfter: [
-        {
-          action: 'click',
-          selector: 'button.dnb-modal__close-button',
-        },
-      ],
+      simulateAfter: { keypress: 'Escape' },
       screenshotSelector: '.dnb-modal__content', // leave one with the background
       rootClassName: 'hide-page-content',
     })
@@ -64,12 +59,7 @@ describe.each(['ui', 'sbanken'])('Dialog for %s', (themeName) => {
       simulate: 'click',
       simulateSelector:
         '[data-visual-test="dialog-fullscreen"] button:first-of-type',
-      simulateAfter: [
-        {
-          action: 'click',
-          selector: 'button.dnb-modal__close-button',
-        },
-      ],
+      simulateAfter: { keypress: 'Escape' },
       screenshotSelector: '.dnb-modal__content .dnb-scroll-view',
       rootClassName: 'hide-page-content',
     })
@@ -82,12 +72,7 @@ describe.each(['ui', 'sbanken'])('Dialog for %s', (themeName) => {
       simulate: 'click',
       simulateSelector:
         '[data-visual-test="dialog-custom-trigger"] button:first-of-type',
-      simulateAfter: [
-        {
-          action: 'click',
-          selector: 'button.dnb-modal__close-button',
-        },
-      ],
+      simulateAfter: { keypress: 'Escape' },
       screenshotSelector: '.dnb-modal__content .dnb-scroll-view',
       rootClassName: 'hide-page-content',
     })
@@ -100,12 +85,7 @@ describe.each(['ui', 'sbanken'])('Dialog for %s', (themeName) => {
       simulate: 'click',
       simulateSelector:
         '[data-visual-test="full-dialog"] button:first-of-type',
-      simulateAfter: [
-        {
-          action: 'click',
-          selector: 'button.dnb-modal__close-button',
-        },
-      ],
+      simulateAfter: { keypress: 'Escape' },
       screenshotSelector: '.dnb-modal__content .dnb-scroll-view',
       rootClassName: 'hide-page-content',
     })
@@ -118,12 +98,7 @@ describe.each(['ui', 'sbanken'])('Dialog for %s', (themeName) => {
       simulate: 'click',
       simulateSelector:
         '[data-visual-test="dialog-default"] button:first-of-type',
-      simulateAfter: [
-        {
-          action: 'click',
-          selector: 'button.dnb-modal__close-button',
-        },
-      ],
+      simulateAfter: { keypress: 'Escape' },
       screenshotSelector: '.dnb-modal__content .dnb-scroll-view',
       rootClassName: 'hide-page-content',
     })
@@ -136,12 +111,7 @@ describe.each(['ui', 'sbanken'])('Dialog for %s', (themeName) => {
       simulate: 'click',
       simulateSelector:
         '[data-visual-test="dialog-confirm-default"] button:first-of-type',
-      simulateAfter: [
-        {
-          action: 'click',
-          selector: 'div#dnb-modal-root button.dnb-button--primary',
-        },
-      ],
+      simulateAfter: { keypress: 'Escape' },
       screenshotSelector: '.dnb-modal__content .dnb-scroll-view',
       rootClassName: 'hide-page-content',
     })
@@ -154,12 +124,7 @@ describe.each(['ui', 'sbanken'])('Dialog for %s', (themeName) => {
       simulate: 'click',
       simulateSelector:
         '[data-visual-test="dialog-confirm-delete"] button:first-of-type',
-      simulateAfter: [
-        {
-          action: 'click',
-          selector: 'div#dnb-modal-root button.dnb-button--primary',
-        },
-      ],
+      simulateAfter: { keypress: 'Escape' },
       screenshotSelector: '.dnb-modal__content .dnb-scroll-view',
       rootClassName: 'hide-page-content',
     })
@@ -172,12 +137,7 @@ describe.each(['ui', 'sbanken'])('Dialog for %s', (themeName) => {
       simulate: 'click',
       simulateSelector:
         '[data-visual-test="dialog-confirm-loggedout"] button:first-of-type',
-      simulateAfter: [
-        {
-          action: 'click',
-          selector: 'div#dnb-modal-root button.dnb-button--primary',
-        },
-      ],
+      simulateAfter: { keypress: 'Escape' },
       screenshotSelector: '.dnb-modal__content .dnb-scroll-view',
       rootClassName: 'hide-page-content',
     })
@@ -190,12 +150,7 @@ describe.each(['ui', 'sbanken'])('Dialog for %s', (themeName) => {
       simulate: 'click',
       simulateSelector:
         '[data-visual-test="dialog-confirm-cookie"] button:first-of-type',
-      simulateAfter: [
-        {
-          action: 'click',
-          selector: 'div#dnb-modal-root button.dnb-button--primary',
-        },
-      ],
+      simulateAfter: { keypress: 'Escape' },
       screenshotSelector: '.dnb-modal__content .dnb-scroll-view',
       rootClassName: 'hide-page-content',
     })
@@ -208,12 +163,7 @@ describe.each(['ui', 'sbanken'])('Dialog for %s', (themeName) => {
       simulate: 'click',
       simulateSelector:
         '[data-visual-test="dialog-progress-indicator"] button:first-of-type',
-      simulateAfter: [
-        {
-          action: 'click',
-          selector: '.dnb-modal__content',
-        },
-      ],
+      simulateAfter: { keypress: 'Escape' },
       screenshotSelector: '.dnb-modal__content .dnb-scroll-view',
       rootClassName: 'hide-page-content',
     })
@@ -239,12 +189,7 @@ describe.each(['ui', 'sbanken'])('Dialog for %s', (themeName) => {
         simulateSelector:
           '[data-visual-test="dialog-scroll-content"] button:first-of-type',
         waitAfterSimulateSelector: '.dnb-scroll-view',
-        simulateAfter: [
-          {
-            action: 'click',
-            selector: 'div#dnb-modal-root button.dnb-button--primary',
-          },
-        ],
+        simulateAfter: { keypress: 'Escape' },
         screenshotSelector: '.dnb-modal__content .dnb-scroll-view',
         rootClassName: 'hide-page-content',
       })
@@ -258,12 +203,7 @@ describe.each(['ui', 'sbanken'])('Dialog for %s', (themeName) => {
         simulateSelector:
           '[data-visual-test="dialog-scroll-content"] button:first-of-type',
         waitAfterSimulateSelector: '.dnb-scroll-view',
-        simulateAfter: [
-          {
-            action: 'click',
-            selector: 'div#dnb-modal-root button.dnb-button--primary',
-          },
-        ],
+        simulateAfter: { keypress: 'Escape' },
         screenshotSelector: '.dnb-modal__content .dnb-scroll-view',
         rootClassName: ['hide-page-content', 'scroll-to-bottom'],
       })
